Fixes for Dlx in Ast cabinet
- Fixed reversed opamp inputs to the 0.1x gain audio path

- Changed the implementation of the "Y flip" jumper. Now the jumper
selects the original signal or the reversed signal, rather than
grounding the "INV_Y" line.  This results in a more predictable result.
I.e., Moving the jumper will always flip the Y axis, while a jumper that
grounds "INV_Y" will have no effect if the line is grounded elsewhere in
the system.

- Bumped version to 4.0

- Thanks to Michael W. for testing the adapter with an Asteroids Deluxe
PCB in an Asteroids Cabinet.