Updated README and Makefile
- README includes a Makefile-generated link to the PCB rendering,
instead of hard-coding the filename in the README, since the name
changes with the revision number.

Add link to rendering

Replace link to rendering with a copy

README uses version-independent link to image

To avoid changing the README every rev bump, make a copy of the top
rendering without the rev number, and include that copy instead.

It would be preferable to generate an include file calling out the
current rendering, but apparently Github won't allow include files in
the README.

Add assembly notes with connector part numbers